+++ src/joi/options.ts
@@ -1,5 +1,11 @@
+// tslint:disable-next-line: no-implicit-dependencies
 import { JSONSchema4 } from 'json-schema';
 
 export interface SubSchemas {
   [k: string]: JSONSchema4;
+}
+
+export interface Options {
+  subSchemas?: SubSchemas;
+  rootSchema?: JSONSchema4;
 }