+++ src/joi/resolve.ts
@@ -1,71 +1,29 @@
+// tslint:disable-next-line: no-implicit-dependencies
 import { JSONSchema4 } from 'json-schema';
-import { JoiSchema, createJoiItem, JoiAlternatives } from './types';
+import { JoiSchema } from './types';
 import { resolveType } from './resolveType';
-import { SubSchemas } from './options';
-import { join } from 'path';
-
-function resolveAnyOf(schemas: JSONSchema4[]): JoiAlternatives {
-  const joiSchema = createJoiItem('alternatives') as JoiAlternatives;
-  joiSchema.try = schemas.map((v) => {
-    return resolveJSONSchema(v);
-  });
-  return joiSchema;
-}
-
-function resolveOneOf(schemas: JSONSchema4[]): JoiAlternatives {
-  return resolveAnyOf(schemas);
-}
-
-function resolveNot(schema: JSONSchema4): JoiAlternatives {
-  const joiSchema = createJoiItem('alternatives') as JoiAlternatives;
-  joiSchema.not = resolveJSONSchema(schema);
-  return joiSchema;
-}
-
-function resolveReference(ref: string, schema: JSONSchema4, subSchema?: SubSchemas): JSONSchema4 {
-  const id = ref.substr(0, ref.indexOf('#') + 1);
-  const path = ref.substr(ref.indexOf('#') + 1);
-
-  let refSchema;
-
-  if (id && subSchema) {
-    refSchema = subSchema[id] || subSchema[id.substr(0, id.length - 1)];
-  }
-
-  if (!refSchema) {
-    refSchema = schema;
-  }
-
-  let fragment = refSchema;
-  path.split('/').some((p) => {
-    fragment = typeof fragment === 'object' && fragment[p];
-    return fragment === undefined;
-  });
-
-  return fragment;
-}
+import { Options } from './options';
+import { resolveJoiAlternativesSchema } from './alternatives';
+import { resolveReference } from './reference';
 
 // tslint:disable-next-line:naming-convention
-export function resolveJSONSchema(schema: JSONSchema4, subSchema?: SubSchemas): JoiSchema {
+export function resolveJSONSchema(schema: JSONSchema4, options?: Options): JoiSchema {
   if (schema.type) {
-    return resolveType(schema);
+    return resolveType(schema, options);
   }
-  if (schema.anyOf) {
-    return resolveAnyOf(schema.anyOf);
+  if (schema.anyOf || schema.not) {
+    return resolveJoiAlternativesSchema(schema, options);
   }
 
   // TODO schema.allOf
 
-  if (schema.oneOf) {
-    return resolveOneOf(schema.oneOf);
-  }
+  // TODO schema.oneof
 
-  if (schema.not) {
-    return resolveNot(schema.not);
-  }
-
-  if (schema.$ref) {
-    return resolveJSONSchema(resolveReference(schema.$ref, schema, subSchema));
+  if (schema.$ref && options) {
+    const ref = resolveReference(schema.$ref, options);
+    if (ref) {
+      return resolveJSONSchema(ref, options);
+    }
   }
 
   if (schema.enum) {